requirements.txt
neo4j
pandas
python-dotenv
fastapi
pydantic
uvicorn
openai